This code has RAM!
This code can remember the location of one ball. You can adjust the "smoothness" of your data by increasing `smoothLevel`, but turn it too high and it wont notice if a ball leaves view for a while. I found 10 was a good number, but that number may change once I add compatibility with multiple balls